Compile bignum tests in both gmp and native modes. Fix some bugs with the native version
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@5212 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ cc-lib/bignum/big.cc
@@ -122,7 +122,7 @@ static constexpr int FIRST_OMITTED_PRIME = 7927;
 #ifndef BIG_USE_GMP
 
 std::vector<std::pair<BigInt, int>>
-BigInt::PrimeFactorization(const BigInt &x, int mf) {
+BigInt::PrimeFactorization(const BigInt &x, int64_t mf) {
   // Simple trial division.
   // It would not be hard to incorporate Fermat's method too,
   // for cases that the number has factors close to its square
